requirements.txt
jinja2>=3.1.0
requests>=2.31.0
htmlmin>=0.1.12
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
responses>=0.23.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
